Metric: Log extended smp pairing events
Include transport and complete information in extended enums.

Tag: #refactor
Bug: 236677941
Test: atest gd/cert/run

+++ system/main/shim/metrics_api.cc
@@ -89,7 +89,7 @@ void LogMetricReadTxPowerLevelResult(const RawAddress& raw_address,
                                                  transmit_power_level);
 }
 
-void LogMetricSmpPairingEvent(const RawAddress& raw_address, uint8_t smp_cmd,
+void LogMetricSmpPairingEvent(const RawAddress& raw_address, uint16_t smp_cmd,
                               android::bluetooth::DirectionEnum direction,
                               uint8_t smp_fail_reason) {
   Address address = bluetooth::ToGdAddress(raw_address);
